data load refactor
