--- src/cms/api/data-load-setting/models/data-load-setting.js
@@ -1,8 +1,0 @@
-'use strict';
-
-/**
- * Read the documentation (https://strapi.io/documentation/developer-docs/latest/development/backend-customization.html#lifecycle-hooks)
- * to customize this model
- */
-
-module.exports = {};
