+++ src/cms/api/x-data-load-setting/config/routes.json
@@ -2,48 +2,48 @@
   "routes": [
     {
       "method": "GET",
-      "path": "/data-load-settings",
-      "handler": "data-load-setting.find",
+      "path": "/x-data-load-settings",
+      "handler": "x-data-load-setting.find",
       "config": {
         "policies": []
       }
     },
     {
       "method": "GET",
-      "path": "/data-load-settings/count",
-      "handler": "data-load-setting.count",
+      "path": "/x-data-load-settings/count",
+      "handler": "x-data-load-setting.count",
       "config": {
         "policies": []
       }
     },
     {
       "method": "GET",
-      "path": "/data-load-settings/:id",
-      "handler": "data-load-setting.findOne",
+      "path": "/x-data-load-settings/:id",
+      "handler": "x-data-load-setting.findOne",
       "config": {
         "policies": []
       }
     },
     {
       "method": "POST",
-      "path": "/data-load-settings",
-      "handler": "data-load-setting.create",
+      "path": "/x-data-load-settings",
+      "handler": "x-data-load-setting.create",
       "config": {
         "policies": []
       }
     },
     {
       "method": "PUT",
-      "path": "/data-load-settings/:id",
-      "handler": "data-load-setting.update",
+      "path": "/x-data-load-settings/:id",
+      "handler": "x-data-load-setting.update",
       "config": {
         "policies": []
       }
     },
     {
       "method": "DELETE",
-      "path": "/data-load-settings/:id",
-      "handler": "data-load-setting.delete",
+      "path": "/x-data-load-settings/:id",
+      "handler": "x-data-load-setting.delete",
       "config": {
         "policies": []
       }
